add load test for syslog receiver
